Added parser logic to parse CAIDA data into TopologySource to later on converted into Graphs and other types for engine

+++ scripts/get-caida.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
+# Make data directory if it doesn't exist
+if [ ! -d data ]; then
+    mkdir data
+fi
+
 # URL of the CAIDA AS relationships data
 BASE_URL="https://publicdata.caida.org/datasets/as-relationships/serial-2/"
 
-# Download the index.html file
-wget -q -O index.html $BASE_URL
+# Download if the the index.html file don't exists
+if [ ! -f index.html ]; then
+    echo "Downloading index.html"
+    wget -q -O index.html $BASE_URL
+fi
 
 # Extract download links from index.html
 grep -oP '(?<=href=")[^"]*\.bz2(?=")' index.html | while read -r FILE; do
@@ -17,8 +25,11 @@ grep -oP '(?<=href=")[^"]*\.bz2(?=")' index.html | while read -r FILE; do
     FILE_URL="${BASE_URL}${FILE}"
     
     # Download the file
-    wget -q $FILE_URL
+    wget $FILE_URL
 done
+
+# Go back to root project directory
+cd ../..
 
 # Clean up
 rm index.html
